Create the initial chaincode and README
This creates the basic chaincode and has been tested using the peer CLI.

It creates a permissioned blob storage using the fabric chaincode
interface.  The user who creates a blob is the only user authorized to
update or delete that blob, though all users may read it.  See the
README.md for more information.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>